Fix failing tests and compilation errors in userspace/bloom
- Fixed math logic in `circle_to_path` which expects 6 VIR segments.
- Corrected the `evicts_when_over_budget` RenderState test to properly exceed the cache capacity and trigger an eviction.
- Repaired `test_damage_collapse_on_overflow` to correctly space out rectangles and updated `add_rect_with_cause` to actively enforce the 25% screen area collapse rule when rects are merged.
- Resolved various compilation and API issues (`error[E0596]`, `error[E0308]`, `UiNodeKind::Crown` mapping, etc).
- Removed duplicate `mod tests;` declarations.
