+++ userspace/bloom/src/vir/mod.rs
@@ -75,39 +75,3 @@ impl Default for VirDocument {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_create_empty_document() {
-        let doc = VirDocument::new();
-        assert_eq!(doc.elements.len(), 0);
-        assert!(doc.width.is_none());
-        assert!(doc.height.is_none());
-        assert!(doc.view_box.is_none());
-    }
-
-    #[test]
-    fn test_create_simple_rect() {
-        let mut path = VirPath::new();
-        path.move_to(0.0, 0.0);
-        path.line_to(100.0, 0.0);
-        path.line_to(100.0, 100.0);
-        path.line_to(0.0, 100.0);
-        path.close();
-
-        let elem = VirElement {
-            path: Arc::new(path),
-            fill: Some(FillStyle {
-                paint: Paint::Solid(VirColor::rgb(255, 0, 0)),
-                rule: FillRule::NonZero,
-            }),
-            stroke: None,
-            transform: VirTransform::identity(),
-            opacity: 1.0,
-        };
-
-        assert_eq!(elem.path.segments.len(), 5);
-    }
-}
